Better handling of pjax
When navigating between the Conversation, Commits, and Files changes
tabs on a pull request, GitHub loads the contents via pjax, updating the
existing DOM instead of doing a full page refresh. The current
implementation of this extension loses the click event bindings after a
pjax page load occurs.

This change fixes that, so that click events are tracked at the document
level and filtered to links we care about. Also includes some
refactoring.