Drastically simplify
Instead of hooking observers up to changes for multiple elements and
trying to tell when certain parts of the page are updated, this observes
the main element and decorates files anytime it changes. This may prove
to be Too Much™, but so far seems to work well.

A side effect is that collapsing a file, navigating to another tab, then
coming back the file is no longer collapsed. I was planning on using the
local store to keep this info anyway, so file state is persisted across
full page loads, so I'm not concerned the happenstance state no longer
holds.