Persist collapsed files
Uses local (extension) storage to track which files on which pages have
been collapsed. This information is used to automatically collapse files
when revisiting a previously viewed page.

Introduces a Store object to encapsulate interactions with the local
storage. An unexpected behaviour of the local storage, is that fetching
a key will return an object _including_ that key, whose value is the
value you actually wanted. E.g.

```
chrome.storage.local.get("myKey", data => {
 // data is { "myKey": "stuff I care about" }
})
```

The `_deserialize` "private" method deals with this. Is there a more
appropriate name for what it's doing?